% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTSE.R
\name{buildTSE}
\alias{buildTSE}
\title{Builds a TreeSummarizedExperiment (TSE) object from salmon quantified files and
file of trees given by TreeTerminus. Internally it reads the salmon
quantified files using tximeta. The \code{rowTree} is the tree on a transcript set.
The transcript set is computed by taking a union of all transcripts that are
covered by the trees from TreeTerminus and transcripts that pass the filtering
criteria determined using \code{labelKeep} function in \code{fishpond}. The
rows consists of all the nodes including the leaf and inner nodes of the
\code{rowTree}. The assays consist of (counts, abundance and infRep matrices)
with the inner node value of an individual assay representing an aggregation
of the descendant leaves for that node in that assay. It requires that Salmon
should have been run with either bootstrap or gibbs sampling.}
\usage{
buildTSE(treeTermFile, coldata, txps = NULL, ...)
}
\arguments{
\item{treeTermFile}{path to file of trees given by TreeTerminus}

\item{coldata}{data.frame that is given as an input to tximeta}

\item{txps}{(Optional) A character vector containing transcripts to which
TreeSummarizedExperiment object will be restricted.}

\item{...}{arguments passed to \code{labelKeep} function in \code{fishpond}}
}
\value{
TreeSummarizedExperiment Object
}
\description{
Builds a TreeSummarizedExperiment (TSE) object from salmon quantified files and
file of trees given by TreeTerminus. Internally it reads the salmon
quantified files using tximeta. The \code{rowTree} is the tree on a transcript set.
The transcript set is computed by taking a union of all transcripts that are
covered by the trees from TreeTerminus and transcripts that pass the filtering
criteria determined using \code{labelKeep} function in \code{fishpond}. The
rows consists of all the nodes including the leaf and inner nodes of the
\code{rowTree}. The assays consist of (counts, abundance and infRep matrices)
with the inner node value of an individual assay representing an aggregation
of the descendant leaves for that node in that assay. It requires that Salmon
should have been run with either bootstrap or gibbs sampling.
}
\examples{
# path to example data
dir <- system.file('extdata/brain_sim_nodtu_small_example', package='beaveR')
# path to file output by TreeTerminus
clustFile <- file.path(dir, 'cluster_nwk.txt')
# path to Salmon quantified files
quantDir <- file.path(dir, 'out_sal')
samples <- as.vector(outer(c(1:6), c(1,2), function(x,y) paste(x,y,sep='_')))
quantFiles <- file.path(quantDir, samples, 'quant.sf')
coldata <- data.frame(files=quantFiles, names=samples, condition=factor(rep(1:2, each=6)))
tse <- buildTSE(treeTermFile = clustFile, coldata = coldata)
}
