% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-metric.R
\name{getLog2FC}
\alias{getLog2FC}
\title{https://github.com/mikelove/fishpond/blob/c1e968f810f2caba2b333a3891d2c3c05e73001b/R/swish.R#L338}
\usage{
getLog2FC(infRepsArray, condition, pc = 5)
}
\description{
https://github.com/mikelove/fishpond/blob/c1e968f810f2caba2b333a3891d2c3c05e73001b/R/swish.R#L338
}
