% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-metric.R
\name{getScaledLFC}
\alias{getScaledLFC}
\title{Computes log fold change on the scaled inferential replicate counts.
Scaled count for a given transcript in an inferential replicate is given by
dividing the original count by the sum of counts of all transcripts in that
inferential replicate}
\usage{
getScaledLFC(tse, x, pc = 5)
}
\arguments{
\item{tse}{TreeSumarizedExperiment obtained as the output of running
\code{buildTSE}}

\item{x}{the name of the condition variable. A factor with two levels for a
two group analysis. The log fold change is computed as non-reference level
over reference level}

\item{pc}{integer pseudocount for computing log fold change}
}
\value{
numeric vector of length equal to number of rows in tse
}
\description{
Computes log fold change on the scaled inferential replicate counts.
Scaled count for a given transcript in an inferential replicate is given by
dividing the original count by the sum of counts of all transcripts in that
inferential replicate
}
\examples{
example(buildTSE)
lfc <- getScaledLFC(tse, 'condition')
}
